requirements.txt
adjusttext==0.8
altair==5.3.0
antropy==0.1.6
anyio==4.4.0
appnope==0.1.4
arch==7.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
category-encoders==2.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmdstanpy==1.2.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
cython==3.0.10
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
ephem==4.1.5
executing==2.0.1
fastjsonschema==2.19.1
flask==3.0.3
fonttools==4.52.1
fqdn==1.5.1
gitdb==4.0.11
gitpython==3.1.43
h11==0.14.0
holidays==0.33
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imbalanced-learn==0.12.3
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipeadatapy==0.1.9
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
isoduration==20.11.0
itables==2.1.1
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.3.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.0
jupyter-server-terminals==0.5.3
jupyterlab==4.2.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.2
jupyterlab-widgets==3.0.11
kaleido==0.2.1
kiwisolver==1.4.5
lightgbm==4.3.0
llvmlite==0.42.0
lunarcalendar==0.0.9
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mizani==0.9.3
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.0
notebook-shim==0.2.4
numba==0.59.1
numpy==1.26.4
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.1.4
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
plotly==5.22.0
plotly-resampler==0.10.0
plotnine==0.12.4
pmdarima==2.0.4
polars==0.20.29
pox==0.3.4
ppft==1.7.6.8
prometheus-client==0.20.0
prompt-toolkit==3.0.45
prophet==1.1.5
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycaret==3.3.2
pycparser==2.22
pydeck==0.9.1
pygments==2.18.0
pyod==1.1.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytimetk==0.4.0
pytz==2024.1
pyyaml==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
send2trash==1.8.3
setuptools==70.0.0
six==1.16.0
sktime==0.26.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
stanio==0.5.0
statsmodels==0.14.2
stochastic==0.6.0
streamlit==1.35.0
supersmoother==0.4
tbats==1.1.3
tenacity==8.3.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
tsdownsample==0.1.3
tsfeatures==0.4.5
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wurlitzer==3.1.0
xarray==2023.10.1
xgboost==2.1.0
xxhash==3.4.1
yellowbrick==1.5
zipp==3.19.0
